--- a/build/optimized.wat
+++ b/build/optimized.wat
@@ -3204,7 +3204,7 @@
   local.get $13
   f64.ne
   if (result i32)
-   i32.const 255
+   i32.const 0
   else   
    local.get $13
    i32.trunc_f64_u