… build/untouched.wat
… package.json
… test/bench.js
… test/test.js