Before (node 12.7.0):
Total Running Time: 708 ms

After:
Total Running Time: 643 ms

Also I found [this regression](https://github.com/AssemblyScript/assemblyscript/issues/748) which could speedup slightly more. But I guess main reason is interop.

Need father investigation.

**PS** Also I switched to `f64` from `f32` because it no produce significant speed gain. Also fixed tests preparation.

**PPS** I excluded diffPtr part and measure only two `wasmModule.pixelmatch` and allocs [here](https://github.com/JamesMilnerUK/pixelmatch-asc/pull/2/commits/b93e9536542edd6ad136c3c6763e17253c8f40a5)

wasm total result:
242 ms

js total result (original mapbox/pixelmatch):
143 ms

JS benchmark looks like (modify test.js):
```js
// eslint-disable-next-line no-unused-vars
function diffTest(imgPath1, imgPath2, diffPath, options, expectedMismatch) {
    // const name = `comparing ${imgPath1} to ${imgPath2}, ${JSON.stringify(options)}`;

    // test(name, (t) => {
    const img1 = readImage(imgPath1);
    const img2 = readImage(imgPath2);
    const {width, height} = img1;
    const diff = new PNG({width, height});

    const start = performance.now();

    match(img1.data, img2.data, diff.data, width, height, options);
    match(img1.data, img2.data, null, width, height, options);

    const end = performance.now();
    total += end - start;

    writeImage(diffPath, diff);
    console.log('\n', imgPath1, imgPath2, end - start, 'ms \n');

    // if (process.env.UPDATE) {
    //     writeImage(diffPath, diff);
    // } else {
    //     const expectedDiff = readImage(diffPath);
    //     t.ok(diff.data.equals(expectedDiff.data), 'diff image');
    // }
    // t.equal(mismatch, expectedMismatch, 'number of mismatched pixels');
    // t.equal(mismatch, mismatch2, 'number of mismatched pixels without diff');

    // t.end();
    // });
}
```