@MaxGraey 2019-08-06
Also you used `unoptimized` wasm for bench here:
https://github.com/JamesMilnerUK/pixelmatch-asc/pull/2/files#diff-3899b83a0e044cf174c10553299b5137R8