@MaxGraey 2019-08-07
For avoid this issue and optimize access to buffers I useed lower level api in some places and this reduce time from `242 ms` to `77` ms which approx twice faster than in JS bench
```
 1a 1b 15.093281000852585 ms 
 2a 2b 11.829204000532627 ms 
 3a 3b 6.862983003258705 ms 
 4a 4b 32.566536001861095 ms 
 5a 5b 3.3966749981045723 ms 
 6a 6b 5.452500998973846 ms 
 6a 6a 2.0666389986872673 ms 

Total Running Time: 77.2678190022707
```