@jtenner 2019-08-07
Is there a reason why you aren't using `memory.compare()` here?