@JamesMilnerUK 2019-08-08
@MaxGraey super, thanks so much for this. There's so styling issues but I'm happy to merge and tidy those up myself. I'd love to hear more about the things that you did for the performance boost too.