@MaxGraey 2019-08-10
Sure.
What had the greatest performance impact:
- Move from high level access to `Uint8Array` to lower level operations `load`/`store` with raw references (`.dataStart`). That's currently workaround due to [this issue](https://github.com/AssemblyScript/assemblyscript/issues/748). Also group sequence RGBA byte access to packed/unpacked u32 access [here](https://github.com/JamesMilnerUK/pixelmatch-asc/pull/2/files#diff-702c9d97e3fad1cceb7d23848770c8b3R269), [here](https://github.com/JamesMilnerUK/pixelmatch-asc/pull/2/files#diff-702c9d97e3fad1cceb7d23848770c8b3R277) and [here](https://github.com/JamesMilnerUK/pixelmatch-asc/pull/2/files#diff-702c9d97e3fad1cceb7d23848770c8b3R277)  which reduce number of read to memory by 4 times.

What had the minor performance impact:
- Use improvements like `if (i32(a == 0) | i32(b == 0))` instead `if (a == 0 || b == 0)` inside loops. AS currently can't properly optimize such logical expressions so we can do this manually. Of course you could safely skip this manual optimizations for clarity and and the speed will not suffer much.

- Optimize `a1 /= 255.0` to `a1 *= 1.0 / 255`. Also micro optimization which usually turned on in LLVM/GCC when `--ffast-math` flag set but binaryen doesn't support that optimization because it slightly decrease float point accuracy but in current application that's not big deal.

Rest of changes just cleanups, code style improvements and simplifications.