@JamesMilnerUK 2019-08-11
@MaxGraey appreciate the  clarification on your performance improvements. This is really epic and cool that it is even faster than the JavaScript implementation. Thank you for all your hard work!